fix(console): load environments in transitions hooks
Fixes gravitee-io/issues#4816
